cleaned the front page, added new /styret page and prepared footer
