+++ src/App.js
@@ -1,11 +1,13 @@
 import React from 'react';
 import Routes from "./App-routes";
 import Header from "./components/header"
+import Footer from "./components/footer"
 
 const App = () => (
   <>
     <Header />
     <Routes />
+    <Footer />
   </>
 );
